+++ resources/js/bootstrap/cp.ts
@@ -23,7 +23,6 @@ import AppLayout from '@/common/layouts/AppLayout.vue';
 import {createCpComponentRegistry} from './components.js';
 import {registerFormComponents} from '@/modules/forms/register';
 import {configureIcons} from './icons.js';
-import LocalFsSettings from '@/components/Filesystems/LocalFsSettings.vue';
 
 let bootedCallbacks: Array<(instance: any) => void> = [];
 let bootingCallbacks: Array<(instance: any) => void> = [];
@@ -155,7 +154,6 @@ const Cp = {
         app.component('AssetIndexes', AssetIndexes);
         app.component('SystemMessages', SystemMessages);
         app.component('CpLink', CpLink);
-        app.component('LocalFsSettings', LocalFsSettings);
 
         components.install(app);
       },